refactor login
