+++ server/model/user.go
@@ -1,8 +1,15 @@
 package model
 
+import "golang.org/x/crypto/bcrypt"
+
 type UserInfo struct {
 	Id   uint   `json:"id"`
 	Name string `json:"name"`
+}
+
+type TryLoginUser struct {
+	Name     string `json:"name"`
+	Password string `json:"password"`
 }
 
 func FindAllUser(users *[]UserInfo) {
@@ -14,4 +21,21 @@ func FindAllUser(users *[]UserInfo) {
 		*users = append(*users, user)
 	}
 	defer db.Close()
+}
+
+func CheckNameAndPassword(tryLoginUser TryLoginUser, findUser *User) (err error) {
+	// dbに接続
+	db := DbConnect()
+	// 名前から名前とパスワードを取得
+	err = db.Where("name = ?", tryLoginUser.Name).First(findUser).Error
+	if err != nil {
+		return err
+	}
+	defer db.Close()
+	// パスワードが一致するか確認
+	err = bcrypt.CompareHashAndPassword(findUser.Password, []byte(tryLoginUser.Password))
+	if err != nil {
+		return err
+	}
+	return nil
 }
